Add critical fixes and comprehensive Sentinel tests
Critical Fixes:
- Fix nil pointer dereference in publisher.Close() (main.go)
- Add int32 to int64 cast for ObservedGeneration compatibility

Test Improvements:
- Add comprehensive unit tests for Sentinel.trigger()
  * TestTrigger_Success: validates event publishing
  * TestTrigger_NoEventsPublished: tests skip logic
  * TestTrigger_FetchError: error handling
  * TestTrigger_PublishError: graceful degradation
  * TestTrigger_MixedResources: complex scenarios
- Sentinel coverage: 0% → 69.8%

All tests passing:
- client: 83.1% coverage
- config: 88.7% coverage
- engine: 100% coverage
- sentinel: 69.8% coverage

+++ internal/client/client.go
@@ -224,7 +224,7 @@ func (c *HyperFleetClient) fetchResourcesOnce(ctx context.Context, _ ResourceTyp
 				Phase:              item.Status.GetPhase(),
 				LastTransitionTime: item.Status.GetLastTransitionTime(),
 				LastUpdated:        item.Status.GetLastUpdated(),
-				ObservedGeneration: item.Status.GetObservedGeneration(),
+				ObservedGeneration: int64(item.Status.GetObservedGeneration()),
 			},
 			Metadata: item.GetMetadata(),
 		}
